+++ java/src/se/sveaekonomi/webpay/integration/order/row/InvoiceFeeBuilder.java
@@ -7,9 +7,9 @@ public class InvoiceFeeBuilder implements RowBuilder {
     private Double amountExVat;
     private Double amountIncVat;
     private Double vatPercent;
-    private int quantity;
+    private Double quantity;
     private String unit;
-    private Integer discountPercent;
+    private Double discountPercent;
     
     public String getName() {
         return name;
@@ -54,7 +54,7 @@ public class InvoiceFeeBuilder implements RowBuilder {
         return this;
     }
     
-    public int getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
     
@@ -88,7 +88,7 @@ public class InvoiceFeeBuilder implements RowBuilder {
         return this;
     }
     
-    public Integer getDiscountPercent() {
+    public Double getDiscountPercent() {
         return discountPercent;
     }
     
@@ -97,7 +97,7 @@ public class InvoiceFeeBuilder implements RowBuilder {
      * @param discountPercent
      * @return InvoiceFeeBuilder
      */
-    public InvoiceFeeBuilder setDiscountPercent(int discountPercent) {
+    public InvoiceFeeBuilder setDiscountPercent(Double discountPercent) {
         this.discountPercent = discountPercent;
         return this;
     }
